middleware/requirements.txt
django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2